REST folder restructure & cleanup
