+++ client/lcd/root.go
@@ -18,9 +18,9 @@ import (
 	tx "github.com/cosmos/cosmos-sdk/client/tx"
 	version "github.com/cosmos/cosmos-sdk/version"
 	"github.com/cosmos/cosmos-sdk/wire"
-	auth "github.com/cosmos/cosmos-sdk/x/auth/rest"
-	bank "github.com/cosmos/cosmos-sdk/x/bank/rest"
-	ibc "github.com/cosmos/cosmos-sdk/x/ibc/rest"
+	auth "github.com/cosmos/cosmos-sdk/x/auth/client/rest"
+	bank "github.com/cosmos/cosmos-sdk/x/bank/client/rest"
+	ibc "github.com/cosmos/cosmos-sdk/x/ibc/client/rest"
 )
 
 const (
