+++ x/bank/client/rest/sendtx.go
@@ -2,6 +2,7 @@ package rest
 
 import (
 	"encoding/hex"
+	"encoding/json"
 	"io/ioutil"
 	"net/http"
 
@@ -11,41 +12,41 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/context"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire"
-	"github.com/cosmos/cosmos-sdk/x/ibc"
+	"github.com/cosmos/cosmos-sdk/x/bank/client"
 )
 
 // RegisterRoutes - Central function to define routes that get registered by the main application
 func RegisterRoutes(r *mux.Router, cdc *wire.Codec, kb keys.Keybase) {
-	r.HandleFunc("/ibc/{destchain}/{address}/send", TransferRequestHandler(cdc, kb)).Methods("POST")
+	ctx := context.NewCoreContextFromViper()
+
+	r.HandleFunc("/accounts/{address}/send", SendRequestHandler(cdc, kb, ctx)).Methods("POST")
 }
 
-type transferBody struct {
+type sendBody struct {
+	// fees is not used currently
 	// Fees             sdk.Coin  `json="fees"`
 	Amount           sdk.Coins `json:"amount"`
 	LocalAccountName string    `json:"name"`
 	Password         string    `json:"password"`
-	SrcChainID       string    `json:"src_chain_id"`
+	ChainID          string    `json:"chain_id"`
 	Sequence         int64     `json:"sequence"`
 }
 
-// TransferRequestHandler - http request handler to transfer coins to a address
-// on a different chain via IBC
-func TransferRequestHandler(cdc *wire.Codec, kb keys.Keybase) func(http.ResponseWriter, *http.Request) {
-	ctx := context.NewCoreContextFromViper()
+// SendRequestHandler - http request handler to send coins to a address
+func SendRequestHandler(cdc *wire.Codec, kb keys.Keybase, ctx context.CoreContext) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		// collect data
 		vars := mux.Vars(r)
-		destChainID := vars["destchain"]
 		address := vars["address"]
 
-		var m transferBody
+		var m sendBody
 		body, err := ioutil.ReadAll(r.Body)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
 			return
 		}
-		err = cdc.UnmarshalJSON(body, &m)
+		err = json.Unmarshal(body, &m)
 		if err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			w.Write([]byte(err.Error()))
@@ -68,8 +69,12 @@ func TransferRequestHandler(cdc *wire.Codec, kb keys.Keybase) func(http.Response
 		to := sdk.Address(bz)
 
 		// build message
-		packet := ibc.NewIBCPacket(info.PubKey.Address(), to, m.Amount, m.SrcChainID, destChainID)
-		msg := ibc.IBCTransferMsg{packet}
+		msg := client.BuildMsg(info.PubKey.Address(), to, m.Amount)
+		if err != nil { // XXX rechecking same error ?
+			w.WriteHeader(http.StatusInternalServerError)
+			w.Write([]byte(err.Error()))
+			return
+		}
 
 		// sign
 		ctx = ctx.WithSequence(m.Sequence)
@@ -88,7 +93,7 @@ func TransferRequestHandler(cdc *wire.Codec, kb keys.Keybase) func(http.Response
 			return
 		}
 
-		output, err := cdc.MarshalJSON(res)
+		output, err := json.MarshalIndent(res, "", "  ")
 		if err != nil {
 			w.WriteHeader(http.StatusInternalServerError)
 			w.Write([]byte(err.Error()))
